Add code to recover previous library path
