+++ python/main-classic/platformcode/xbmc_helpers.py
@@ -1,0 +1,71 @@
+import xbmc
+from core import filetools
+from core import logger
+
+
+def execute_sql_kodi(sql):
+    """
+    Ejecuta la consulta sql contra la base de datos de kodi
+    @param sql: Consulta sql valida
+    @type sql: str
+    @return: Numero de registros modificados o devueltos por la consulta
+    @rtype nun_records: int
+    @return: lista con el resultado de la consulta
+    @rtype records: list of tuples
+    """
+    logger.info()
+    file_db = ""
+    nun_records = 0
+    records = None
+
+    # Buscamos el nombre de la BBDD de videos segun la version de kodi
+    code_db = {'10': 'MyVideos37.db', '11': 'MyVideos60.db', '12': 'MyVideos75.db', '13': 'MyVideos78.db',
+               '14': 'MyVideos90.db', '15': 'MyVideos93.db', '16': 'MyVideos99.db', '17': 'MyVideos107.db'}
+
+    video_db = code_db.get(xbmc.getInfoLabel("System.BuildVersion").split(".", 1)[0], '')
+    if video_db:
+        file_db = filetools.join(xbmc.translatePath("special://userdata/Database"), video_db)
+
+    # metodo alternativo para localizar la BBDD
+    if not file_db or not filetools.exists(file_db):
+        file_db = ""
+        for f in filetools.listdir(xbmc.translatePath("special://userdata/Database")):
+            path_f = filetools.join(xbmc.translatePath("special://userdata/Database"), f)
+
+            if filetools.isfile(path_f) and f.lower().startswith('myvideos') and f.lower().endswith('.db'):
+                file_db = path_f
+                break
+
+    if file_db:
+        logger.info("Archivo de BD: %s" % file_db)
+        conn = None
+        try:
+            import sqlite3
+            conn = sqlite3.connect(file_db)
+            cursor = conn.cursor()
+
+            logger.info("Ejecutando sql: %s" % sql)
+            cursor.execute(sql)
+            conn.commit()
+
+            records = cursor.fetchall()
+            if sql.lower().startswith("select"):
+                nun_records = len(records)
+                if nun_records == 1 and records[0][0] is None:
+                    nun_records = 0
+                    records = []
+            else:
+                nun_records = conn.total_changes
+
+            conn.close()
+            logger.info("Consulta ejecutada. Registros: %s" % nun_records)
+
+        except:
+            logger.error("Error al ejecutar la consulta sql")
+            if conn:
+                conn.close()
+
+    else:
+        logger.debug("Base de datos no encontrada")
+
+    return nun_records, records
